+++ core/triggers/interface_99_moddiscountrules_discountrulesTriggers.class.php
@@ -130,6 +130,32 @@ class InterfaceDiscountrulesTriggers extends DolibarrTriggers
 
 			$line = $currentObject;
 
+			if($line->element == 'facturedet'){
+				/** @var FactureLigne $line  */
+				$parentObject = new Facture($line->db);
+				$resFetchParent = $parentObject->fetch($line->fk_facture);
+			}
+			elseif($line->element == 'commandedet'){
+				/** @var OrderLine $line */
+				$parentObject = new Commande($line->db);
+				$resFetchParent = $parentObject->fetch($line->fk_commande);
+			}
+			elseif($line->element == 'propaldet'){
+				/** @var PropaleLigne $line */
+				$parentObject = new Propal($line->db);
+				$resFetchParent = $parentObject->fetch($line->fk_propal);
+			}
+			else{
+				// UNKNOW ELEMENT OR NOT COMPATIBLE
+				return 0;
+			}
+
+			if($resFetchParent < 0){
+				$this->errors[] = 'Error fetching parent document for discount rules';
+				return -1;
+			}
+
+			//Only on propal cause error during conversion
 			if(getDolGlobalInt('DISCOUNTRULES_USE_MARKUP_MARGIN_RATE') && $action == 'LINEPROPAL_INSERT' && !getDolGlobalString('MAIN_ROUNDOFTOTAL_NOT_TOTALOFROUND')){
 
 					//Define value for conf DISCOUNTRULES_MARKUP_MARGIN_RATE & DISCOUNTRULES_MINIMUM_RATE
@@ -142,28 +168,28 @@ class InterfaceDiscountrulesTriggers extends DolibarrTriggers
 					$minimumRate = getDolGlobalInt('DISCOUNTRULES_MINIMUM_RATE');
 
 					if (!empty($valueConfMarkupMarginRate)){
-						$costPrice = 0;
-						if(empty($line->fk_product)){
-							$costPrice = GETPOST('buying_price', 'int');
-							if (empty($costPrice)){
-								return 0;
-							}
+						$costPrice = GETPOST('buying_price', 'int');
 
-						} else {
-							$myProduct = new Product($this->db);
-							$myProduct->fetch($currentObject->fk_product);
-							$costPrice = $myProduct->cost_price;
+						$myProduct = new Product($this->db);
+						$myProduct->fetch($currentObject->fk_product);
+						$parentObject->fetch_thirdparty();
+						if (! empty($parentObject->thirdparty->array_options['options_discountrules_min_markup_margin_percent'])) {
+							$minimumRate = $parentObject->thirdparty->array_options['options_discountrules_min_markup_margin_percent'];
+						} else if (! empty($myProduct->array_options['options_discountrules_min_markup_margin_percent'])) {
+							$minimumRate = $myProduct->array_options['options_discountrules_min_markup_margin_percent'];
 						}
+						$costPrice = $currentObject->pa_ht ?? $myProduct->cost_price;
+
 
 						if ($costPrice) {
 							//Subprice for MarkRate / MarginRate
 							if ($valueConfMarkupMarginRate == 'MarkRate' ){
-								$currentObject->subprice = price2num($costPrice / (1 - $minimumRate / 100));
+								$currentObject->subprice = price2num($costPrice / (1 - $minimumRate / 100), 'MT');
 							}elseif ($valueConfMarkupMarginRate == 'MarginRate'){
-								$currentObject->subprice = price2num($costPrice * (1 + $minimumRate / 100));
+								$currentObject->subprice = price2num($costPrice * (1 + $minimumRate / 100), 'MT');
 							}
 							//Total HT
-							$currentObject->total_ht = $currentObject->subprice * $currentObject->qty;
+							$currentObject->total_ht = price2num($currentObject->subprice * $currentObject->qty, 'MT');
 							$res = $currentObject->update($user);
 							if(!$res){
 								setEventMessages($currentObject->error, $currentObject->errors, 'errors');
@@ -194,31 +220,6 @@ class InterfaceDiscountrulesTriggers extends DolibarrTriggers
 
 			if($forceUpdateDiscount){
 				dol_syslog("Trigger '".$this->name."' for action '$action' launched by ".__FILE__.". id=".$line->id);
-
-				if($line->element == 'facturedet'){
-					/** @var FactureLigne $line  */
-					$parentObject = new Facture($line->db);
-					$resFetchParent = $parentObject->fetch($line->fk_facture);
-				}
-				elseif($line->element == 'commandedet'){
-					/** @var OrderLine $line */
-					$parentObject = new Commande($line->db);
-					$resFetchParent = $parentObject->fetch($line->fk_commande);
-				}
-				elseif($line->element == 'propaldet'){
-					/** @var PropaleLigne $line */
-					$parentObject = new Propal($line->db);
-					$resFetchParent = $parentObject->fetch($line->fk_propal);
-				}
-				else{
-					// UNKNOW ELEMENT OR NOT COMPATIBLE
-					return 0;
-				}
-
-				if($resFetchParent < 0){
-					$this->errors[] = 'Error fetching parent document for discount rules';
-					return -1;
-				}
 
 				$dateTocheck = time();
 				if (!getDolGlobalInt('DISCOUNTRULES_SEARCH_WITHOUT_DOCUMENTS_DATE')) $dateTocheck = $parentObject->date;