+++ ChangeLog.md
@@ -1,9 +1,13 @@
 # CHANGELOG FOR DISCOUNTRULES MODULE
 
 
-
-
 ## Not Released
+- NEW : Min markup/margin extrafield now shows on customer third parties (wrong display condition on Societe object) - *05/12/2025* - 2.28.0
+
+## RELEASE 2.27
+- NEW : Minimum markup rate by customer or product + refacto in js file - *23/09/2025* - 2.27.0
+
+## RELEASE 2.26
 - NEW : COMPAT V22 - *08/07/2025* - 2.26.0
 
 ## RELEASE 2.25